massive commit

+++ libft/srcs/ft_mapnget.c
@@ -6,7 +6,7 @@
 /*   By: olkovale <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/22 10:10:12 by olkovale          #+#    #+#             */
-/*   Updated: 2017/09/21 02:23:57 by olkovale         ###   ########.fr       */
+/*   Updated: 2017/09/23 21:42:04 by olkovale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include "libft.h"
 
 t_map_kv	*ft_mapnget(t_map *mp, void *key,
-						int (*keycmp)(void *, void *, size_t), int sz)
+						int (*keycmp)(void *, void *, int), int sz)
 {
 	t_map_kv	*iter;
 	t_map_kv	*end;
@@ -26,7 +26,7 @@ t_map_kv	*ft_mapnget(t_map *mp, void *key,
 	end = iter + mp->sz;
 	while (iter < end)
 	{
-		if (0 == keycmp(key, iter->key, (size_t)sz))
+		if (0 == keycmp(key, iter->key, sz))
 			return (iter);
 		iter++;
 	}
